Move linker to the acme package.
