+++ acme/linker_test.go
@@ -1,4 +1,4 @@
-package api
+package acme
 
 import (
 	"context"
